Throw exception if no gallery tiles found.
Within MosaicStepFunctions, in the DetermineBestImages step, a list of available tiles is needed. The LoadGalleryItems uses a tablename and a gallery name to return a list of tiles available for mosaic construction. Previously, an empty list of tiles could be returned from this function. Because it is not useful to return an empty list from LoadGalleryItems, the function now throws an exception if no tiles are found.

The primary motivator for this change is better error messages and faster debugging. If an empty list of tiles is used an IndexOutOfRangeException will occur later in the program.